+++ itsf_sales_coding_test_6/src/main/product/MusicCD.java
@@ -9,7 +9,7 @@ public class MusicCD extends Product {
 
 	private static final ProductType MUSIC_CD_PRODUCT_TYPE = ProductType.Other;
 	
-	public MusicCD(Integer price, Packaging packaging, Boolean isImported) {
+	public MusicCD(Float price, Packaging packaging, Boolean isImported) {
 		super(price, MUSIC_CD_PRODUCT_TYPE, packaging, isImported);
 	}
