+++ itsf_sales_coding_test_6/src/main/product/Perfume.java
@@ -9,7 +9,7 @@ public class Perfume extends Product {
 
 	private static final ProductType PERFUME_PRODUCT_TYPE = ProductType.Other;
 	
-	public Perfume(Integer price, Packaging packaging, Boolean isImported) {
+	public Perfume(Float price, Packaging packaging, Boolean isImported) {
 		super(price, PERFUME_PRODUCT_TYPE, packaging, isImported);
 	}
